\name{CNpcgdp}
\alias{CNpcgdp}
\docType{data}
\title{Provincial Per Capita GDP in China, 2016--2024}
\description{
Annual per capita GDP for 31 provincial-level regions in China from 2016 to 2024.
}
\usage{data(CNpcgdp)}
\format{
  A data frame with 31 observations on the following 10 variables:
  \describe{
    \item{\code{province}}{Provincial-level region name.}
    \item{\code{pcgdp2016}}{Per capita GDP in 2016 (CNY per person).}
    \item{\code{pcgdp2017}}{Per capita GDP in 2017 (CNY per person).}
    \item{\code{pcgdp2018}}{Per capita GDP in 2018 (CNY per person).}
    \item{\code{pcgdp2019}}{Per capita GDP in 2019 (CNY per person).}
    \item{\code{pcgdp2020}}{Per capita GDP in 2020 (CNY per person).}
    \item{\code{pcgdp2021}}{Per capita GDP in 2021 (CNY per person).}
    \item{\code{pcgdp2022}}{Per capita GDP in 2022 (CNY per person).}
    \item{\code{pcgdp2023}}{Per capita GDP in 2023 (CNY per person).}
    \item{\code{pcgdp2024}}{Per capita GDP in 2024 (CNY per person).}
  }
}
\source{
  National Bureau of Statistics of China (NBS), data portal:
  \url{https://data.stats.gov.cn/}
}
\keyword{datasets}