\name{Energy}
\alias{Energy}
\docType{data}
\title{Selected Global Energy Indicators}
\description{
Annual energy-related indicators for selected countries.
}
\usage{data(Energy)}
\format{
A data frame containing:
\describe{
  \item{\code{country}}{Country name.}
  \item{\code{year}}{Year.}
  \item{\code{pop}}{Population.}
  \item{\code{gdp}}{Gross domestic product.}
  \item{\code{energy}}{Primary energy consumption.}
  \item{\code{elec}}{Electricity generation.}
  \item{\code{renew}}{Share of renewable energy.}
}
}
\source{
Our World in Data energy dataset.
\url{https://github.com/owid/energy-data}
}
\keyword{datasets}