% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hare}
\alias{Hare}
\title{Hare Population Data}
\format{
A time series object with the following characteristics:
\itemize{
\item Time period: 1845 - 1935
\item Frequency: Annual
\item Values: Hare population counts
}
}
\source{
Package 'astsa' available at \href{https://nickpoison.github.io/}{https://nickpoison.github.io/}
}
\usage{
data(Hare)
}
\description{
Hare population data from the book "AI enabled time series and spacial statistics"
}
\details{
Hare population data

This dataset contains hare population data from the book "AI enabled time series and spacial statistics".

The Hare dataset provides annual population counts of hares from 1845 to 1935. This dataset is often used in time series analysis and population dynamics studies.
}
\examples{
data(Hare)

# Plot the hare population data
plot(Hare, main = "Hare Population Data", 
     xlab = "Year", ylab = "Population", 
     col = "blue")

# Calculate summary statistics
summary(Hare)
}
\keyword{datasets}
