% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lap}
\alias{lap}
\title{Local Area Pollution (LAP) Data}
\format{
A multivariate time series (mts) object with the following characteristics:
\itemize{
\item Time period: 1970 - 1980
\item Frequency: Weekly (52 observations per year)
\item Number of variables: 11
\item Variables:
\itemize{
\item tmort: Total mortality
\item rmort: Respiratory mortality
\item cmort: Cardiovascular mortality
\item tempr: Temperature
\item rh: Relative humidity
\item co: Carbon monoxide
\item so2: Sulfur dioxide
\item no2: Nitrogen dioxide
\item hycarb: Hydrocarbons
\item o3: Ozone
\item part: Particulate matter
}
}
}
\source{
Package 'astsa' available at \href{https://nickpoison.github.io/}{https://nickpoison.github.io/}
}
\usage{
data(lap)
}
\description{
LAP data from the book "AI enabled time series and spacial statistics"
}
\details{
LAP data

This dataset contains LAP (Local Area Pollution) data from the book "AI enabled time series and spacial statistics".

The lap dataset provides weekly measurements of various pollution and health indicators from 1970 to 1980.
This dataset is often used in environmental health studies and time series analysis of pollution effects.
}
\examples{
data(lap)

# Plot the first variable (total mortality)
plot(lap[, 1], main = "Total Mortality", 
     xlab = "Time", ylab = "Mortality Rate", 
     col = "red")

# Calculate correlation between temperature and ozone
cor(lap[, "tempr"], lap[, "o3"], use = "complete.obs")

# Plot multiple variables
plot(lap[, c("tempr", "o3", "co")], 
     main = "Temperature, Ozone, and Carbon Monoxide")
}
\keyword{datasets}
