% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lynx}
\alias{Lynx}
\title{Lynx Population Data}
\format{
A time series object with the following characteristics:
\itemize{
\item Time period: 1845 - 1935
\item Frequency: Annual
\item Values: Lynx population counts
}
}
\source{
Package 'astsa' available at \href{https://nickpoison.github.io/}{https://nickpoison.github.io/}
}
\usage{
data(Lynx)
}
\description{
Lynx population data from the book "AI enabled time series and spacial statistics"
}
\details{
Lynx population data

This dataset contains lynx population data from the book "AI enabled time series and spacial statistics".

The Lynx dataset provides annual population counts of lynx from 1845 to 1935. This dataset is often used in time series analysis and predator-prey relationship studies.
}
\examples{
data(Lynx)

# Plot the lynx population data
plot(Lynx, main = "Lynx Population Data", 
     xlab = "Year", ylab = "Population", 
     col = "red")

# Calculate summary statistics
summary(Lynx)
}
\keyword{datasets}
