\name{WorldMacro}
\alias{WorldMacro}
\docType{data}
\title{Global Macroeconomic Panel Data}
\description{
Panel data of macroeconomic indicators for multiple countries,
including GDP per capita, population, and inflation rate.
}
\usage{data(WorldMacro)}
\format{
A data frame containing:
\describe{
  \item{\code{country}}{Country name.}
  \item{\code{year}}{Year.}
  \item{\code{gdp_pc}}{GDP per capita (constant USD).}
  \item{\code{population}}{Total population.}
  \item{\code{inflation}}{Inflation rate.}
}
}
\source{
World Bank World Development Indicators
\url{https://data.worldbank.org}
}
\keyword{datasets}