% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hare}
\alias{Hare}
\title{Snowshoe Hare Population Data}
\format{
A time series object with the following characteristics:
\itemize{
\item Time period: 1845 - 1935
\item Frequency: Annual
\item Values: Number of snowshoe hare pelts traded (in thousands)
}
}
\source{
Package 'astsa' available at \href{https://nickpoison.github.io/}{https://nickpoison.github.io/}
}
\usage{
data(Hare)
}
\description{
Snowshoe hare population data from the Hudson's Bay Company records, covering the period from 1845 to 1935. This classic dataset is widely used for studying predator-prey population dynamics.
}
\details{
Hare population data

This dataset contains snowshoe hare population data collected by the Hudson's Bay Company.

The Hare dataset provides annual counts of snowshoe hare pelts traded by the Hudson's Bay Company from 1845 to 1935. This dataset, along with the Lynx dataset, forms a classic example of predator-prey population dynamics and is commonly used in time series analysis and ecological modeling.
}
\examples{
data(Hare)

# Plot the hare population data
plot(Hare, main = "Snowshoe Hare Population Data", 
     xlab = "Year", ylab = "Pelts Traded (thousands)", 
     col = "blue")

# Calculate summary statistics
summary(Hare)
}
\keyword{datasets}
