% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lynx}
\alias{Lynx}
\title{Canadian Lynx Population Data}
\format{
A time series object with the following characteristics:
\itemize{
\item Time period: 1845 - 1935
\item Frequency: Annual
\item Values: Number of lynx pelts traded (in thousands)
}
}
\source{
Package 'astsa' available at \href{https://nickpoison.github.io/}{https://nickpoison.github.io/}
}
\usage{
data(Lynx)
}
\description{
Canadian lynx population data from the Hudson's Bay Company records, covering the period from 1845 to 1935. This classic dataset is widely used for studying predator-prey population dynamics alongside the Hare dataset.
}
\details{
Lynx population data

This dataset contains Canadian lynx population data collected by the Hudson's Bay Company.

The Lynx dataset provides annual counts of Canadian lynx pelts traded by the Hudson's Bay Company from 1845 to 1935. Together with the Hare dataset, this forms a classic example of predator-prey population dynamics, where lynx (predators) and snowshoe hares (prey) exhibit cyclical population fluctuations.
}
\examples{
data(Lynx)

# Plot the lynx population data
plot(Lynx, main = "Canadian Lynx Population Data", 
     xlab = "Year", ylab = "Pelts Traded (thousands)", 
     col = "red")

# Calculate summary statistics
summary(Lynx)

# Plot both hare and lynx populations together
plot(Hare, main = "Hare and Lynx Population Dynamics", 
     xlab = "Year", ylab = "Pelts Traded (thousands)", 
     col = "blue", ylim = c(0, max(Hare, Lynx)))
lines(Lynx, col = "red")
legend("topleft", legend = c("Hare", "Lynx"), 
       col = c("blue", "red"), lty = 1)
}
\keyword{datasets}
